requirements.txt
django==4.2.1
mysqlclient==2.2.7
python-dotenv
